🐛  Remove static methods from SMS class
Type(s): FIX

Why:
SMS class should not contain statics. As it sits on the container, if
static is used then container can be bypassed, and it will pollute the
global space. And also, changing provider on the fly does not work.
Static methods are provided via the facade itself.

What:
Make SMS class accesible via instance. Change alias in composer.json, as
it should point to the facade.